+++ src/Vortice.DXGI/IDXGIDevice.cs
@@ -8,26 +8,6 @@ namespace Vortice.DXGI
 {
     public partial class IDXGIDevice
     {
-        /// <inheritdoc/>
-        protected override unsafe void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                ReleaseDevice();
-            }
-            base.Dispose(disposing);
-        }
-
-        private void ReleaseDevice()
-        {
-            if (Adapter__ != null)
-            {
-                // Don't use Dispose() in order to avoid circular references
-                ((IUnknown)Adapter__).Release();
-                Adapter__ = null;
-            }
-        }
-
         public IDXGISurface CreateSurface(IntPtr sharedResource)
         {
             if (sharedResource == IntPtr.Zero)