+++ src/Vortice.Direct3D11/D3D11.cs
@@ -10,14 +10,11 @@ namespace Vortice.Direct3D11
 {
     public static partial class D3D11
     {
-        public static Result D3D11CreateDevice(IDXGIAdapter adapter,
-            DriverType driverType,
-            DeviceCreationFlags flags,
-            FeatureLevel[] featureLevels,
+        public static Result D3D11CreateDevice(IDXGIAdapter adapter, DriverType driverType, DeviceCreationFlags flags, FeatureLevel[] featureLevels,
             out ID3D11Device device)
         {
             return RawD3D11CreateDeviceNoContext(
-                adapter,
+                adapter != null ? adapter.NativePointer : IntPtr.Zero,
                 driverType,
                 flags,
                 featureLevels,
@@ -25,25 +22,19 @@ namespace Vortice.Direct3D11
                 out _);
         }
 
-        public static Result D3D11CreateDevice(IDXGIAdapter adapter,
-            DriverType driverType,
-            DeviceCreationFlags flags,
-            FeatureLevel[] featureLevels,
-            out ID3D11Device device,
-            out ID3D11DeviceContext immediateContext)
+        public static Result D3D11CreateDevice(IDXGIAdapter adapter, DriverType driverType, DeviceCreationFlags flags, FeatureLevel[] featureLevels,
+            out ID3D11Device device, out ID3D11DeviceContext immediateContext)
         {
             return D3D11CreateDevice(adapter, driverType, flags, featureLevels, out device, out _, out immediateContext);
         }
 
-        public static Result D3D11CreateDevice(IDXGIAdapter adapter,
-            DriverType driverType,
-            DeviceCreationFlags flags,
-            FeatureLevel[] featureLevels,
-            out ID3D11Device device,
-            out FeatureLevel featureLevel,
-            out ID3D11DeviceContext immediateContext)
+        public static Result D3D11CreateDevice(IDXGIAdapter adapter, DriverType driverType, DeviceCreationFlags flags, FeatureLevel[] featureLevels,
+            out ID3D11Device device, out FeatureLevel featureLevel, out ID3D11DeviceContext immediateContext)
         {
-            Result result = D3D11CreateDevice(adapter, driverType, IntPtr.Zero,
+            Result result = D3D11CreateDevice(
+                adapter != null ? adapter.NativePointer : IntPtr.Zero,
+                driverType,
+                IntPtr.Zero,
                 (int)flags,
                 featureLevels,
                 (featureLevels != null) ? featureLevels.Length : 0,
@@ -67,7 +58,52 @@ namespace Vortice.Direct3D11
             return result;
         }
 
-        public unsafe static Result D3D11On12CreateDevice(
+        public static Result D3D11CreateDevice(IntPtr adapterPtr, DriverType driverType, DeviceCreationFlags flags, FeatureLevel[] featureLevels,
+            out ID3D11Device device)
+        {
+            return RawD3D11CreateDeviceNoContext(
+                adapterPtr,
+                driverType,
+                flags,
+                featureLevels,
+                out device,
+                out _);
+        }
+
+        public static Result D3D11CreateDevice(IntPtr adapterPtr, DriverType driverType, DeviceCreationFlags flags, FeatureLevel[] featureLevels,
+            out ID3D11Device device, out ID3D11DeviceContext immediateContext)
+        {
+            return D3D11CreateDevice(adapterPtr, driverType, flags, featureLevels, out device, out _, out immediateContext);
+        }
+
+        public static Result D3D11CreateDevice(IntPtr adapterPtr, DriverType driverType, DeviceCreationFlags flags, FeatureLevel[] featureLevels,
+            out ID3D11Device device, out FeatureLevel featureLevel, out ID3D11DeviceContext immediateContext)
+        {
+            Result result = D3D11CreateDevice(adapterPtr, driverType, IntPtr.Zero,
+                (int)flags,
+                featureLevels,
+                (featureLevels != null) ? featureLevels.Length : 0,
+                SdkVersion,
+                out device,
+                out featureLevel,
+                out immediateContext);
+
+            if (result.Failure)
+            {
+                return result;
+            }
+
+            if (immediateContext != null)
+            {
+                device.AddRef();
+                device.ImmediateContext__ = immediateContext;
+                immediateContext.Device__ = device;
+            }
+
+            return result;
+        }
+
+        public static unsafe Result D3D11On12CreateDevice(
             IUnknown d3d12Device,
             DeviceCreationFlags flags,
             FeatureLevel[] featureLevels,
@@ -77,7 +113,7 @@ namespace Vortice.Direct3D11
             out ID3D11DeviceContext immediateContext,
             out FeatureLevel chosenFeatureLevel)
         {
-            var result = D3D11On12CreateDevice(d3d12Device,
+            Result result = D3D11On12CreateDevice(d3d12Device,
                 flags,
                 featureLevels, featureLevels.Length,
                 commandQueues, commandQueues.Length,
@@ -107,12 +143,12 @@ namespace Vortice.Direct3D11
         /// <returns><c>true</c> if the primary adapter is supporting this feature level; otherwise, <c>false</c>.</returns>
         public static bool IsSupportedFeatureLevel(FeatureLevel featureLevel, DeviceCreationFlags flags = DeviceCreationFlags.None)
         {
-            var result = RawD3D11CreateDeviceNoDeviceAndContext(
-                null,
+            Result result = RawD3D11CreateDeviceNoDeviceAndContext(
+                IntPtr.Zero,
                 DriverType.Hardware,
                 flags,
                 new[] { featureLevel },
-                out var outputLevel);
+                out FeatureLevel outputLevel);
             return result.Success && outputLevel == featureLevel;
         }
 
@@ -131,12 +167,36 @@ namespace Vortice.Direct3D11
             if (adapter == null)
                 throw new ArgumentNullException(nameof(adapter), "Invalid adapter");
 
-            var result = RawD3D11CreateDeviceNoDeviceAndContext(
-                adapter,
+            Result result = RawD3D11CreateDeviceNoDeviceAndContext(
+                adapter.NativePointer,
                 DriverType.Unknown,
                 flags,
                 new[] { featureLevel },
-                out var outputLevel);
+                out FeatureLevel outputLevel);
+            return result.Success && outputLevel == featureLevel;
+        }
+
+        /// <summary>
+        /// Check if a feature level is supported by a particular adapter.
+        /// </summary>
+        /// <param name="adapterPtr">The native handle of <see cref="IDXGIAdapter"/>.</param>
+        /// <param name="featureLevel">The feature level.</param>
+        /// <param name="flags">The <see cref="DeviceCreationFlags"/> flags.</param>
+        /// <returns><c>true</c> if the specified adapter is supporting this feature level; otherwise, <c>false</c>.</returns>
+        public static unsafe bool IsSupportedFeatureLevel(
+            IntPtr adapterPtr,
+            FeatureLevel featureLevel,
+            DeviceCreationFlags flags = DeviceCreationFlags.None)
+        {
+            if (adapterPtr == IntPtr.Zero)
+                throw new ArgumentNullException(nameof(adapterPtr), "Invalid adapter handle");
+
+            Result result = RawD3D11CreateDeviceNoDeviceAndContext(
+                adapterPtr,
+                DriverType.Unknown,
+                flags,
+                new[] { featureLevel },
+                out FeatureLevel outputLevel);
             return result.Success && outputLevel == featureLevel;
         }
 
@@ -147,11 +207,11 @@ namespace Vortice.Direct3D11
         public static FeatureLevel GetSupportedFeatureLevel()
         {
             RawD3D11CreateDeviceNoDeviceAndContext(
-                null,
+                IntPtr.Zero,
                 DriverType.Hardware,
                 DeviceCreationFlags.None,
                 null,
-                out var featureLevel);
+                out FeatureLevel featureLevel);
             return featureLevel;
         }
 
@@ -166,11 +226,30 @@ namespace Vortice.Direct3D11
                 throw new ArgumentNullException(nameof(adapter), "Invalid adapter");
 
             RawD3D11CreateDeviceNoDeviceAndContext(
-                adapter,
+                adapter.NativePointer,
                 DriverType.Unknown,
                 DeviceCreationFlags.None,
                 null,
-                out var featureLevel);
+                out FeatureLevel featureLevel);
+            return featureLevel;
+        }
+
+        /// <summary>
+        /// Gets the highest supported hardware feature level of the primary adapter.
+        /// </summary>
+        /// <param name="adapterPtr">The native handle of <see cref="IDXGIAdapter"/>.</param>
+        /// <returns>The highest supported hardware feature level.</returns>
+        public static FeatureLevel GetSupportedFeatureLevel(IntPtr adapterPtr)
+        {
+            if (adapterPtr == IntPtr.Zero)
+                throw new ArgumentNullException(nameof(adapterPtr), "Invalid adapter handle");
+
+            RawD3D11CreateDeviceNoDeviceAndContext(
+                adapterPtr,
+                DriverType.Unknown,
+                DeviceCreationFlags.None,
+                null,
+                out FeatureLevel featureLevel);
             return featureLevel;
         }
 
@@ -183,18 +262,18 @@ namespace Vortice.Direct3D11
             unsafe
             {
                 Result result = D3D11CreateDevice_(
-                        null, (int)DriverType.Null,
-                        null, (int)DeviceCreationFlags.Debug,
-                        null, 0,
-                        SdkVersion,
-                        null, null, null);
+                    null, (int)DriverType.Null,
+                    null, (int)DeviceCreationFlags.Debug,
+                    null, 0,
+                    SdkVersion,
+                    null, null, null);
                 return result.Success;
             }
         }
 
         #region RawD3D11CreateDevice Methods
         private static Result RawD3D11CreateDeviceNoContext(
-            IDXGIAdapter adapter,
+            IntPtr adapterPtr,
             DriverType driverType,
             DeviceCreationFlags flags,
             FeatureLevel[] featureLevels,
@@ -204,11 +283,10 @@ namespace Vortice.Direct3D11
             unsafe
             {
                 device = default;
-                var adapterPtr = CppObject.ToCallbackPtr<IDXGIAdapter>(adapter);
                 fixed (void* featureLevelsPtr = &featureLevels[0])
                 fixed (void* featureLevelPtr = &featureLevel)
                 {
-                    var devicePtr = IntPtr.Zero;
+                    IntPtr devicePtr = IntPtr.Zero;
                     Result result = D3D11CreateDevice_(
                         (void*)adapterPtr,
                         (int)driverType,
@@ -225,15 +303,13 @@ namespace Vortice.Direct3D11
                     {
                         device = new ID3D11Device(devicePtr);
                     }
-
-                    GC.KeepAlive(adapter);
                     return result;
                 }
             }
         }
 
         private static Result RawD3D11CreateDeviceNoDeviceAndContext(
-            IDXGIAdapter adapter,
+            IntPtr adapterPtr,
             DriverType driverType,
             DeviceCreationFlags flags,
             FeatureLevel[] featureLevels,
@@ -241,7 +317,6 @@ namespace Vortice.Direct3D11
         {
             unsafe
             {
-                var adapterPtr = CppObject.ToCallbackPtr<IDXGIAdapter>(adapter);
                 fixed (void* featureLevelsPtr = &featureLevels[0])
                 fixed (void* featureLevelPtr = &featureLevel)
                 {
@@ -256,7 +331,6 @@ namespace Vortice.Direct3D11
                         null,
                         featureLevelPtr,
                         null);
-                    GC.KeepAlive(adapter);
                     return result;
                 }
             }