+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-Current Stable Release:    1.8.6 (August 2020)
+Current Stable Release:    1.8.9 (August 2020)
 
 -----------------------------------------------
 Release:     1.8.x (XXX 2020)
@@ -9,6 +9,9 @@ KEY CHANGES:
 - [FIX] D3D11: ID3D11DeviceContext fix various calls and avoid usage of params that can lead memory allocation.
 - [ENH] D3D11: ID3D11Device - Add feature check helper methods.
 - [ENH] D3D11: ID3D11DeviceContext Add VS/HS/DS/GS/PS/CS UnsetConstantBuffer faster method and fix allocation in ID3D11CommandList.
+- [FIX] DXGI: IDXGIDevice::GetAdapter is not property and not persisted, user need to manually Dispose the IDXGIAdapter (just like native code).
+- [ENH] D3D11: D3D11CreateDevice takes also native IntPtr adapter.
+- [ENH] Samples: Cleanup code and remove all leakages.
 
 -----------------------------------------------
 Release:     1.7.37 (August 2020)